requirements.txt
asgiref==3.2.10
astroid==2.4.2       
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20   
chardet==3.0.4       
click==7.1.2
colorama==0.4.3
Django==3.1
futures==3.1.1
goslate==1.5.1
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
PyDictionary==2.0.1
pylint==2.5.3
python-datamuse==1.3.0
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
wrapt==1.12.1